surya: Add 4MB overhead space to super partition
As per AOSP documentation:

For non-A/B devices and retrofit A/B devices, the sum of maximum sizes of
all groups must be: BOARD_SUPER_PARTITION_SIZE - overhead

Overhead is required in the computation to account for metadata, alignments
and so on. A reasonable overhead is 4 MiB, but you can pick a larger
overhead as needed by the device.

Upon checking CAF, it is noted that they are also using AOSP's recommended
value of 4MB overhead space [2] [3]. Hence, let's do the same.

[1]: https://source.android.com/devices/tech/ota/dynamic_partitions/implement#board-configuration-changes
[2]: https://source.codeaurora.org/quic/la/platform/vendor/qcom/steppe/tree/BoardConfig.mk?h=LA.UM.9.1.r1-09100-SMxxx0.0#n117
[3]: https://source.codeaurora.org/quic/la/platform/vendor/qcom/steppe/tree/BoardConfig.mk?h=LA.UM.9.1.r1-09100-SMxxx0.0#n124